pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-aviator"
version = "1.0.0"
description = "Pilot the world's first electric aircraft — a CLI + Pygame hybrid game"
readme = "readme.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Vlad Beresnev"},
]
keywords = ["game", "pygame", "aviation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Games/Entertainment",
]
dependencies = [
    "mysql-connector-python>=9.0",
    "pygame>=2.6",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
]

[project.scripts]
the-aviator = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
